define patch module
