+++ lib/appium_lib/core/ios.rb
@@ -1,6 +1,6 @@
 # loaded in common/driver.rb
 require_relative 'ios/search_context'
 require_relative 'ios/device'
-require_relative 'ios/patch'
 
+require_relative 'ios/uiautomation/patch'
 require_relative 'ios/uiautomation/bridge'
