+++ lib/appium_lib/core/ios_xcuitest.rb
@@ -1,7 +1,7 @@
 # loaded in common/driver.rb
 require_relative 'ios/search_context'
 require_relative 'ios/device'
-require_relative 'ios/patch'
+require_relative 'ios/uiautomation/patch'
 
 require_relative 'ios/xcuitest/search_context'
 require_relative 'ios/xcuitest/device'
